mythscreenwizard: rename commandlineparser.* to mythscreenwizard_…
